Add CurrencyRatesViewModel ### -Create CurrencyRatesViewModelTest -Create required data classes CurrencyRateViewData -Define Extensions -Follow Red-Green-Refactor TDD cycle

+++ app/src/main/java/com/k1apps/currencyrates/domain/usecase/CurrencyRateData.kt
@@ -1,3 +1,3 @@
 package com.k1apps.currencyrates.domain.usecase
 
-data class CurrencyRateData(val symbol: String, val price: Float)
+data class CurrencyRateData(val symbol: String, val price: Double)
